fixing route management, fix answer time in mathblocks game, and create card game color blocks coming soon
