+++ domains/Game/data.js
@@ -1,10 +1,16 @@
 export const games = [
   {
+    href: "/game/math-blocks",
     title: "Math Blocks",
     description: `
       Kamu diberi waktu untuk menghafal letak angka-angka pada block 2 x 2, 3 x 3,
       dst selama beberapa detik. Setelah itu akan muncul pertanyaan matematika yang
       jawabannya ada di antara block block tersebut dan kamu harus memilih angka 
       dengan jawaban yang tepat.`,
+  },
+  {
+    href: "/game/color-blocks",
+    title: "Color Blocks",
+    description: `Coming soon...`,
   },
 ];
