+++ domains/Game/sections/ColorBlocks/data.js
@@ -1,0 +1,4 @@
+export const data = {
+  title: "Color Blocks",
+  description: `Coming soon...`,
+};
