+++ domains/Game/sections/ColorBlocks/style.js
