+++ domains/Game/sections/MathBlocks/data.js
@@ -1,3 +1,12 @@
+export const data = {
+  title: "Math Blocks",
+  description: `
+    Kamu diberi waktu untuk menghafal letak angka-angka pada block 2 x 2, 3 x 3,
+    dst selama beberapa detik. Setelah itu akan muncul pertanyaan matematika yang
+    jawabannya ada di antara block block tersebut dan kamu harus memilih angka 
+    dengan jawaban yang tepat.`,
+};
+
 export const levels = {
   1: {
     rows: 2,
