+++ domains/Game/style.js
@@ -4,9 +4,13 @@ export const Container = styled.div`
   display: flex;
   flex-wrap: wrap;
 
-  > div.card {
-    margin-bottom: 1rem;
-    text-align: center;
-    user-select: none;
+  > a {
+    width: 100%;
+
+    > div.card {
+      margin-bottom: 1rem;
+      text-align: center;
+      user-select: none;
+    }
   }
 `;
